+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Unreleased
 
 - **[NEW]** Add `IllustratedSection` component
+- **[NEW]** Add `LockIcon`
+- **[UPDATE]** Fix `Button` with `Icon` display case
+- [...]
 
 # v22.0.0 (26/02/2020)